+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.5.4 - 2025-03-03
+
+- Re-add filter removal functionality
+- Fix bug in search field
+
 ## 1.5.3 - 2025-02-28
 
 - Live filtering for type and subject